requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Flask-WTF==1.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pymysql==1.1.0
cryptography==41.0.4